fix: int64 indexing migration batch 6 - loop counters and index types
- Default.All.cs: int i -> long i for size iteration
- Default.Any.cs: int i -> long i for size iteration
- StackedMemoryPool.cs: int i -> long i (count param is long)
- np.random.poisson.cs: int i -> long i for size iteration
- np.random.bernoulli.cs: int i -> long i for size iteration
- np.random.randn.cs: int i -> long i for size iteration
- NDArray.Indexing.Masking.cs:
  - int idx -> long idx for trueCount iteration
  - GetInt32 -> GetInt64 (nonzero returns NDArray<long>[])
  - int valueIdx -> long valueIdx for mask.size iteration

All changes follow INT64_DEVELOPER_GUIDE.md patterns.

+++ src/NumSharp.Core/Backends/Unmanaged/Pooling/StackedMemoryPool.cs
@@ -204,7 +204,7 @@ _restart:
                 var block = new UnmanagedMemoryBlock<byte>(blocksize);
 
                 var addr = new IntPtr(block.Address);
-                for (int i = 0; i < count; i++, addr += SingleSize)
+                for (long i = 0; i < count; i++, addr += SingleSize)
                     availables_blocks.Push(addr);
 
                 _blocks.Add(block);
